refactor: host/invitee poll and confirmed event views render on event data change #57
