+++ src/js/components/event/invitee-poll.js
@@ -8,10 +8,10 @@ import Button from '../common/Button';
 import styles from '../../../styles';
 
 const InviteePoll = ({ event, toggleSelection, poll, handleVote, //eslint-disable-line
-  event_id, isHost, hasVoted }) => { //eslint-disable-line
+  event_id, isHost, hasMadeChoice }) => { //eslint-disable-line
 
-    console.log("event", event);
-  const voteButtonText = hasVoted ? 'VOTE AGAIN' : 'VOTE';
+
+  const voteButtonText = hasMadeChoice ? 'VOTE AGAIN' : 'VOTE';
 
   function eventNote (event) {
 
@@ -27,23 +27,10 @@ const InviteePoll = ({ event, toggleSelection, poll, handleVote, //eslint-disabl
 
   return (
     <View style={styles.Poll}>
-      { eventNote(event) }
+      <Text>POLL (INVITEE VIEW)</Text>
 
       <View style={styles.eventWhat}>
-        {
-          event._what.map((choice, i) => {
-            return
-            <EventWhatSection
-              type={ 'what' }
-              index={ i }
-              isOption={ event._what }
-              toggleSelection={ toggleSelection }
-              text={ choice }
-              choiceClasses={ selectedClasses }
-              labelClasses={ hideClasses }
-            />
-          })
-        }
+        <Text>WHAT</Text>
       </View>
       <View style={styles.eventWhere}>
         <Text>WHERE</Text>
@@ -62,46 +49,6 @@ const InviteePoll = ({ event, toggleSelection, poll, handleVote, //eslint-disabl
     </View>
   );
 };
-
-
-function createPollSelections (event, toggleSelection, poll, eventType, EventTypeComponent) {
-
-
-  return event[eventType].map((choice, i) => {
-
-    const selectedClasses = (poll[eventType] === undefined || poll[eventType][i] === true);
-    const hideClasses = i > 0;
-
-    if (poll[eventType]) { // if this category has multiple options
-
-      return (
-        <EventTypeComponent
-          key={ `${eventType}-${i}` }
-          type={ eventType }
-          index={ i }
-          isOption={ !!poll[eventType] }
-          toggleSelection={ toggleSelection }
-          text={ choice }
-          choiceClasses={ selectedClasses }
-          labelClasses={ hideClasses }
-        />
-      );
-    } else { // if this category is firm
-
-      return (
-        <Button
-          key={ `${eventType}-${i}` }
-        >
-          <EventTypeComponent
-            text={ choice }
-            choiceClasses={ selectedClasses }
-            labelClasses={ hideClasses }
-          />
-        </Button>
-      );
-    }
-  });
-}
 
 
 export default InviteePoll;
